Add the ability for computeExternalSigningInputs to take FieldLike and ViewKeylike inputs
