+++ sdk/src/browser.ts
@@ -18,7 +18,6 @@ import { FunctionObject } from "./models/functionObject.js";
 import { ImportedVerifyingKeys, ImportedPrograms } from "./models/imports.js";
 import { InputJSON } from "./models/input/inputJSON.js";
 import { InputObject } from "./models/input/inputObject.js";
-import { ExternalSigningInput, ExternalSigningOptions, RequestSignInput } from "./models/ExternalSigningInputs";
 import { OutputJSON } from "./models/output/outputJSON.js";
 import { OutputObject } from "./models/output/outputObject.js";
 import { OwnedFilter } from "./models/record-scanner/ownedFilter.js";
@@ -238,8 +237,6 @@ export {
     KeyVerifier,
     MemKeyVerifier,
     Metadata,
-    ExternalSigningInput,
-    ExternalSigningOptions,
     NetworkRecordProvider,
     OfflineKeyProvider,
     OfflineSearchParams,
@@ -275,7 +272,6 @@ export {
     RecordScannerOptions,
     RecordNotFoundError,
     RecordScannerRequestError,
-    RequestSignInput,
     ViewKeyNotStoredError,
     RecordSearchParams,
     RegisterResult,
@@ -338,4 +334,7 @@ export type {
     ViewKeyStrategy,
     InputIdStrategy,
     InputStrategy,
+    ExternalSigningInput,
+    ExternalSigningOptions,
+    RequestSignInput,
 } from "./mpc.js";
